chore(query): move insert format deser from binder into interpreter
