+++ src/query/pipeline/sources/src/processors/sources/mod.rs
@@ -20,7 +20,6 @@ pub mod file_splitter;
 pub mod multi_file_splitter;
 mod one_block_source;
 pub mod stream_source;
-pub mod stream_source_v2;
 pub mod sync_source;
 pub mod sync_source_receiver;
 
@@ -37,7 +36,6 @@ pub use multi_file_splitter::OperatorInfo;
 pub use one_block_source::OneBlockSource;
 pub use stream_source::StreamSource;
 pub use stream_source::StreamSourceNoSkipEmpty;
-pub use stream_source_v2::StreamSourceV2;
 pub use sync_source::SyncSource;
 pub use sync_source::SyncSourcer;
 pub use sync_source::*;
