+++ src/query/service/src/pipelines/processors/mod.rs
@@ -35,7 +35,6 @@ pub use sources::FileSplitterState;
 pub use sources::MultiFileSplitter;
 pub use sources::OperatorInfo;
 pub use sources::StreamSource;
-pub use sources::StreamSourceV2;
 pub use sources::SyncSource;
 pub use sources::SyncSourcer;
 pub use transforms::AggregatorParams;
